Cart UI

+++ public_html/Project/sql/007_create_table_cart.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS RM_Cart(
     product_id int,
     desired_quantity int,
     user_id int,
-	unit_price int,
+	unit_price float,
     created TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     modified TIMESTAMP DEFAULT CURRENT_TIMESTAMP on update CURRENT_TIMESTAMP,
     FOREIGN KEY (user_id) REFERENCES Users(id),
